requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2024.1
crispy-bootstrap5==0.7
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2
django-cache-url==3.4.4
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-djconfig==0.11.0
django-haystack==3.2.1
django-infinite-scroll-pagination==1.1.0
django-js-asset==2.1.0
django-spirit==0.13.0
django-widget-tweaks==1.5.0
environs==9.5.0
gunicorn==20.1.0
Jinja2==3.1.3
mysqlclient
Pillow==9.5.0
psycopg==3.1.8
psycopg-binary==3.1.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
Whoosh==2.7.4
